Factored CipherSegmentProcessor out into own module Added object cleanup code to failure path
